refactor dto go swift renderers
